Modules rewrite: CLI, Lexer, Parser + Analyzer
Basically, this project has been turned completely on its head due to
the introduction of modules. This required an extensive rewrite of the
whole stack, mostly due to the fact that many members of type `String`
now map to the new `ModuleRef` type.

Introduced:
- A new CLI that allows the user to test files against different passes
  in cheshire (`lex` -> `parse` -> `ana` -> `tyck` -> `impl` -> `tr`)
- FileRegistry, to centralize references of files
- Analysis passes for modules
- Parsing steps for `use` and `use pub`
- Generalized visitor architecture

Major changes to:
- The AST, since now programs are rooted in `AstProgram`...
- Analysis architecture (`ana`)
  - Specifically, analysis passes are now structured so they can affect
    the `AnalyzedProgram` as well, meaning we have to be far less
    careful about the ordering of the passes (though care must be taken
    in some cases still).
- The way that the lexer and parser deal with file strings, instead of
  a bulky string being lugged around
- The way that errors propagate their file information around: `FileId`

And a lot of other changes that I have yet to document.

Right now, I'm working on rebuilding the analysis pass from scratch.
Since `ModuleRef` pretty much corresponds to `String`, it shouldn't be
too hard to add type-checking and instantiation back into the mix.
Translation shouldn't be too hard to do as well, and now that I have
modules, I should be able to flesh out a standard library and lift
operators into traits. Once I do that, and implement a very basic
runtime, I should be able to actually test executables!

TODO:
- Finish `ana`
- Return and fix up `tyck` and `inst`
- Modify `tr` (specifically decoration) to handle modules
- Write an operator-aware standard library
- Write a basic runtime to implement standard library
- Mock out a super basic runtime
... Woohoo!