[tyck] Fix non-union in static call type and Self type
Given a trait:

```
trait Foo {
    fn bar(self).
}
```

A call such as:

```
let c: C = ...
<A as Foo>::bar(c).
```

would be valid even if type `A` != type `C`. This is because I was never
uniting the self type of a static call with the actual call type (A).
The self type would later be united with `C` as a regular parameter, but
this wouldn't matter because `A` and `C` were never joined in the
middle via this self type. This meant that pretty much any `impl` would
match something... In this case, Chris found out that `'c'.into()` would
match the `Int` to `String` conversion logic, and this was only caught
at _COMPILE TIME_!!!

Yikes.