[tr] [std] Officially finish v1 of GC
Ok. This is the final really beefy commit for the garbage collection
API. I think it's solid!

- Fixed a bug where I wasn't emitting String logic in gc_visit.
- Fixed a bug where I was accidentally not skipping the right things in
gc_visit_array, since you should quit if callback returns false...
- Generalized the gc visit architecture to take a general callback, we
use that to both mark and relocate(+unmark) pointers in the heap and on
the stack.
- Change the visit/mark infrastructure to take a int** instead of an
int*. This was already how gc_visit was doing it, but also generalized
this to mark and gc_visit_array, which is needed for relocation to work.
- Changed all the usages of malloc everywhere.

Implemented:
- gc_alloc_block = replaces malloc
- gc_get_mark = replaces that stupid &MARK logic everywhere.
- gc_remap_*, logic to store and recall remappings from the sweep gc
step. This is right now an associated list, but I could replace this
with a real hashmap if I really cared...
- gc_walk generalizes the mark logic to a general callback.