[AST] Finally derive(Visit) is a macro
I snapped and finally implemented a derive(Visit) procedural macro. I
originally had planned on it being more complex, with the ability to
derive a whole Adapter trait from the and all that craziness, but I took
it down a notch and just implemented the Visit trait, which actually
ends up being most of the pain anyways.